requirements.txt
fastapi~=0.116.1
prometheus_client~=0.22.1
PyJWT~=2.10.1
uvicorn